+++ test/hexapdf/document/test_fonts.rb
@@ -9,7 +9,7 @@ describe HexaPDF::Document::Fonts do
     @doc.config['font_loader'] = []
   end
 
-  describe "load" do
+  describe "add" do
     before do
       @doc.config['font_loader'] << lambda do |doc, name, **options|
         assert_equal(@doc, doc)
@@ -26,23 +26,23 @@ describe HexaPDF::Document::Fonts do
     end
 
     it "loads the specified font" do
-      assert_equal(:NormalFont, @doc.fonts.load(:TestFont).name)
-      assert_equal(:BoldFont, @doc.fonts.load(:TestFont, variant: :bold).name)
+      assert_equal(:NormalFont, @doc.fonts.add(:TestFont).name)
+      assert_equal(:BoldFont, @doc.fonts.add(:TestFont, variant: :bold).name)
     end
 
     it "caches loaded fonts" do
-      font = @doc.fonts.load(:TestFont)
-      assert_same(font, @doc.fonts.load(:TestFont))
-      assert_same(font, @doc.fonts.load(:TestFont, variant: :none))
+      font = @doc.fonts.add(:TestFont)
+      assert_same(font, @doc.fonts.add(:TestFont))
+      assert_same(font, @doc.fonts.add(:TestFont, variant: :none))
     end
 
     it "fails if the requested font is not found"  do
-      assert_raises(HexaPDF::Error) { @doc.fonts.load("Unknown") }
+      assert_raises(HexaPDF::Error) { @doc.fonts.add("Unknown") }
     end
 
     it "raises an error if a font loader cannot be correctly retrieved" do
       @doc.config['font_loader'][0] = 'UnknownFontLoader'
-      assert_raises(HexaPDF::Error) { @doc.fonts.load(:Other) }
+      assert_raises(HexaPDF::Error) { @doc.fonts.add(:Other) }
     end
   end
 end